models moved outside of project
