+++ tsconfig.json
@@ -7,5 +7,6 @@
     "esModuleInterop": true,
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true
-  }
+  },
+  "references": [{ "path": "../shoppingcard-models" }],
 }
